Adding actions to nav graph, finishing bottomNavView, adding fragments
