+++ androidHackathonApp/src/main/java/com/kcc/kmmhackathon/androidHackathonApp/AddMealViewModel.kt
@@ -1,0 +1,7 @@
+package com.kcc.kmmhackathon.androidHackathonApp
+
+import androidx.lifecycle.ViewModel
+
+class AddMealViewModel : ViewModel() {
+    // TODO: Implement the ViewModel
+}
