+++ androidHackathonApp/src/main/java/com/kcc/kmmhackathon/androidHackathonApp/view/MainActivity.kt
@@ -1,0 +1,23 @@
+package com.kcc.kmmhackathon.androidHackathonApp.view
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import androidx.navigation.findNavController
+import androidx.navigation.ui.setupActionBarWithNavController
+import androidx.navigation.ui.setupWithNavController
+import com.google.android.material.bottomnavigation.BottomNavigationView
+import com.kcc.kmmhackathon.androidHackathonApp.R
+
+class MainActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+        setTitle(R.string.app_name)
+
+        val bottomNavigationView: BottomNavigationView = findViewById(R.id.bottom_nav)
+        val navController = findNavController(R.id.fragment)
+
+        bottomNavigationView.setupWithNavController(navController)
+    }
+}
