+++ androidHackathonApp/src/main/res/layout/activity_main.xml
@@ -6,17 +6,17 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@+id/main_view"
-    tools:context=".MainActivity">
+    tools:context=".view.MainActivity">
 
   <fragment
-      android:id="@+id/fragment2"
+      android:id="@+id/fragment"
       android:name="androidx.navigation.fragment.NavHostFragment"
       android:layout_width="match_parent"
-      android:layout_height="match_parent"
+      android:layout_height="0dp"
       app:defaultNavHost="true"
-      app:layout_constraintBottom_toTopOf="@id/bottom_nav"
       app:layout_constraintEnd_toEndOf="parent"
       app:layout_constraintStart_toStartOf="parent"
+      app:layout_constraintBottom_toTopOf="@id/bottom_nav"
       app:layout_constraintTop_toTopOf="parent"
       app:navGraph="@navigation/main_nav_graph" />
 
@@ -31,5 +31,7 @@
       app:itemIconTint="@drawable/bottom_nav_selector"
       app:itemTextColor="@drawable/bottom_nav_selector"
       app:menu="@menu/bottom_nav_menu" />
+
+
 
 </androidx.constraintlayout.widget.ConstraintLayout>
