+++ androidHackathonApp/src/main/res/layout/fragment_maps.xml
@@ -7,5 +7,6 @@
     android:layout_height="match_parent"
     tools:context=".view.MapsFragment" >
 
+
 </fragment>
