+++ androidHackathonApp/src/main/res/menu/bottom_nav_menu.xml
@@ -3,12 +3,12 @@
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <item
-        android:id="@+id/mapFragment"
+        android:id="@+id/mapsFragment"
         android:icon="@drawable/ic_baseline_map_24"
         android:title="@string/map_tab_label" />
 
     <item
-        android:id="@+id/settingsFragment"
+        android:id="@+id/settingsFragment2"
         android:icon="@drawable/ic_baseline_settings_24"
         android:title="@string/settings_tab_label" />
